Issue-151: Addressing PR Comments
+ field names match J2735 ASN1 schema field names
+ Added README file MovementState and MovementEventList description
+ Added RAPIDJSON_HAS_STRING definition into CMake

+++ streets_utils/streets_signal_phase_and_timing/src/models/time_change_details.cpp
@@ -47,7 +47,7 @@ namespace signal_phase_and_timing {
             }
             if (val.FindMember("confidence")->value.IsUint() ) {
                 // OPTIONAL see J2735 TimeChangeDetails definition 
-                confidence = val["confidence"].GetUint();
+                confidence = static_cast<uint8_t>(val["confidence"].GetUint());
             }
         }
     }
